move SHTC1Sensor class to dedicated sensors directory, add sensors as EXTRA_COMPONENT_DIR
