+++ main/CMakeLists.txt
@@ -1,4 +1,4 @@
-idf_component_register(SRCS "SHTC1Sensor.cpp" "main.cpp"
+idf_component_register(SRCS "main.cpp"
                        PRIV_REQUIRES spi_flash driver freertos
-                       REQUIRES shtc1
+                       REQUIRES shtc3
                        INCLUDE_DIRS "")
