+++ sensors/shtc3/CMakeLists.txt
@@ -1,0 +1,5 @@
+idf_component_register(SRCS "SHTC1Sensor.cpp"
+                       PRIV_REQUIRES spi_flash driver
+                       REQUIRES shtc1
+                       INCLUDE_DIRS ""
+                       )
