+++ sensors/shtc3/SHTC1Sensor.h
